AnalogSource: make the Arduino AnalogSource interrupt driven
this fixes several problems with reading analog sources:

 - we were getting poor values because we didn't wait long enough for
   an analog source to settle

 - we wasted a lot of CPU cycles waiting for conversions

 - we were not taking averages over many samples, which we did with
   the old AP_ADC driver on the APM1

+++ libraries/AP_AnalogSource/AP_AnalogSource_ADC.h
@@ -8,13 +8,13 @@
 class AP_AnalogSource_ADC : public AP_AnalogSource
 {
     public:
-    AP_AnalogSource_ADC( AP_ADC * adc, int ch, float prescale = 1.0 ) :
+    AP_AnalogSource_ADC( AP_ADC * adc, uint8_t ch, float prescale = 1.0 ) :
         _adc(adc), _ch(ch), _prescale(prescale) {}
     float read(void);
 
     private:
     AP_ADC * _adc;
-    int _ch;
+    uint8_t _ch;
     float _prescale;
 };
